requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
greenlet==3.2.2
griffe==1.7.3
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
llama-cloud==0.1.21
llama-cloud-services==0.6.15
llama-index==0.12.37
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.37
llama-index-embeddings-openai==0.3.1
llama-index-graph-stores-nebula==0.4.2
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-deepseek==0.1.1
llama-index-llms-openai==0.3.42
llama-index-llms-openai-like==0.3.5
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-milvus==0.8.2
llama-parse==0.6.12
markupsafe==3.0.2
marshmallow==3.26.1
milvus-lite==2.4.12
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
nebula3-python==3.8.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
openai==1.79.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
pydantic==2.11.4
pydantic-core==2.33.2
pyjwt==2.8.0
pymilvus==2.5.9
pyparsing==3.2.3
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.46.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
zhipuai==2.1.5.20250421
